Attempt at default templating
Attempts to introduce default templating.
Meaning admins can specify a default template in the backend, which will
automatically be set and applied to all new events and series.

Current approach is not great, as it causes additional rerenders.
it also does not work for new events,
as it requires setting "				enableReinitialize" on Formik, which causes
formik state to get overwritten all the time.

Assumes there is a key like "default_template=553" in the
etc/listproviders/acl.default.create.properties