Fix ui for new targetFilters structure
The structure for targetFilters changed in the backend.
There is now another map wrapper that splits filters
by metadata catalogs.
This commit makes the necessary code changes to
accomadate the new structure.
Still, the behaviour remains the same, meaning only
filters for the common event metadata can be added
or changed. This does not add support for other catalogs.