Make sure filter values reset when changing lifecycle policy filter type
A filter for a lifecycle policy can have different types, like
"text" or "date". When changing from one type to another type,
the value selectors should change accordingly (i.e. displaying a
date selector for filter of type "date", then displaying a dropdown
for a filter of type "text". This was not working, but should be
working now.